+++ src/pages/Profile.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Profile = () => {
+  return (
+    <div>
+      <p>Hello World</p>
+    </div>
+  );
+};
+
+export default Profile;
